Tuesday changes
** 0_read_me_first.txt:  finalized docs here

** introduction_to_sensors: revised and finalized per Leng's feedback

** analogTempSensor.docx:

The C source code here #includes math.h, so needs a special command
line flag. I changed the documentation to say:

			gcc analogTemp.c -o analogTemp.out –lm –lwiringPi

which I believe is correct, but ***LENG*** can you verify? Can I use
two -l flags side by side for math and wiringPi?

** how to use a breadboard:

First, to edit this, Word made me rename it as .docx rather than .doc.
So filename change.

I changed some of the contents here, eliminating redundancy and also
the description of "Virtual Breadboards" which I don't think are
relevant to the K47 use of them.

I had one technical concern I could not address however. The last
picture is a breadboard circuit with an LED and resistor.  In this
illustration, the long leg of the LED is connected to the black wire of
the battery. Usually the long LED leg is positive (anode), and the
black wire of a battery is negative, so I think this illustratoin is in
error. I have written the text to match the illustration but if I'm
correct about the error, both the illustration and the text should be
updated. LENG can you decide what's best here?

** (everyExperiment).docx:

Made capitalization of headers consistent; changed bodyfonts to
single-spaced; used Times New Roman and Courier consistently throughout.

Changed "READ_ME.TXT" to "READ_ME_FIRST.TXT".